requirements.txt
PyQt6
pandas
Pillow
openpyxl
ezdxf
opencv-python
numpy
PyMuPDF